% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderbook_env.R
\name{modify.orderBookL2}
\alias{modify.orderBookL2}
\title{Update orderBookL2}
\usage{
\method{modify}{orderBookL2}(ob, inc, action = "update",
  strict = FALSE, copy = FALSE)
}
\description{
Update orderBookL2
}
\details{
modify by reference!!
}
