% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderbook_env.R
\name{new_orderBookL2}
\alias{new_orderBookL2}
\title{#' @export
as_tibble.msg = function(msg) {
  msg$data = list(as_tibble(msg$data))
  as_tibble(msg[attr(msg, 'names')])
}}
\usage{
new_orderBookL2()
}
\description{
#' @export
as_tibble.msg = function(msg) {
  msg$data = list(as_tibble(msg$data))
  as_tibble(msg[attr(msg, 'names')])
}
}
