% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_engine.R
\name{ob_increment}
\alias{ob_increment}
\title{Make snapshot from event records
Returns tibble (hash table)}
\usage{
ob_increment(snapshot, increment, key = c("id", "side"))
}
\description{
Make snapshot from event records
Returns tibble (hash table)
}
